demo/examples/NutrionMobileApp/model-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.15.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
